Add converter code (#10)
* Build converter for Alberta XML format to OCDS JSON.

This CL adds the framework for the general conversion facilities along
with adding the specific converter for the Alberta XML data format.

* Add converter for BCBid data.

This CL adds the converter for BCBid data to OCDS JSON.

* Add Nova Scotia to OCDS converter.

This CL adds a converter to go from the Nova Scotia data format to OCDS
JSON.

* Add Nunavut converter.

This CL adds the converter code from a Nunavut CSV dump to OCDS JSON.

* Refactor CSV base class.

This CL moves the common CSV code into a CSVInput class which is then
used by Nunavut and Nova Scotia importers.

* Add Yukon converter

This CL adds a converter for the Yukon CSV to OCDS JSON format.

* Add QC converter.

This CL adds a converter for the SEAO Avis XML file to OCDS.

* Update OCDS converter to new schema.

The CETA OCDS schema was updated to match the GPA schema. This CL
updates the converters to generate the matching OCDS files.

* Add new CETA fields to schema

* Importer CETA schema updates

* Eager load sub-resources on index call

* Set title of post based on description if needed.

If the title is not provided, synthesize one based off the first number
of characters from the description.

* More null checks

This CL adds more checks before outputting entries into the JSON files
to make sure they aren't empty.

* Fixup data for procuring entities.

This CL fixes loading the identifier for procuring entities and the
output of blank names.

* Add Open Canada CSV to OCDS converter.

This CL adds a converter from the Open Canada CSV format to OCDS.

* Add converter tests

This CL adds tests for the converter code. The procurement controller
tests which are no longer needed are removed.

* Minor controller changes.

This CL limits the index page to 25 random entries along with limiting
the search results to 25 records. The search detection of the searchable
type has also been fixed.